Post mixin test refactor. Post #request fulfill/unfulfill tests stubbed in
